dsp-blocks/features-from-time-series-images/requirements-blocks.txt
audioread==2.1.9
librosa==0.8.0
matplotlib==3.5.1
numpy==1.21.6
PeakUtils==1.3.2
Pillow==9.0.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
scikit-learn==0.21.2
scipy==1.7.3
sklearn==0.0
urllib3==1.24.2
PyWavelets==1.3.0